+++ frontend/src/pages/Navbar/ProjectSearch.js
@@ -1,14 +1,7 @@
-import ButtonBase from "@material-ui/core/ButtonBase";
-import FormControl from "@material-ui/core/FormControl";
-import IconButton from "@material-ui/core/IconButton";
-import InputBase from "@material-ui/core/InputBase";
-import Paper from "@material-ui/core/Paper";
-import Tooltip from "@material-ui/core/Tooltip";
-import CancelIcon from "@material-ui/icons/Cancel";
-import SearchIcon from "@material-ui/icons/Search";
 import React from "react";
 import { withStyles } from "@material-ui/core";
-import { DebounceInput } from "react-debounce-input";
+import Searchbar from "./Searchbar";
+import strings from "../../localizeStrings";
 
 const styles = {
   searchField: {
@@ -31,61 +24,16 @@ const ProjectSearch = ({
   storeSearchTerm
 }) => {
   return (
-    <div style={{ display: "flex" }}>
-      {searchBarDisplayed && !searchDisabled ? (
-        <Paper style={styles.searchField}>
-          <form onSubmit={e => e.preventDefault()} style={{ width: "90%" }}>
-            <FormControl style={{ width: "97%", paddingLeft: "5px" }} data-test="project-search-field">
-              <DebounceInput
-                element={InputBase}
-                debounceTimeout={300}
-                onChange={event => storeSearchTerm(event.target.value)}
-                onKeyDown={e => {
-                  if (e.key === "Escape" || e.key === "Esc") {
-                    storeSearchTerm("");
-                    storeSearchBarDisplayed(false);
-                  }
-                }}
-                style={styles.debounceInput}
-                value={searchTerm}
-                autoFocus={true}
-              />
-            </FormControl>
-          </form>
-          <ButtonBase
-            data-test="clear-project-search"
-            onClick={() => {
-              storeSearchBarDisplayed(false);
-              storeSearchTerm("");
-            }}
-          >
-            <CancelIcon color="action" />
-          </ButtonBase>
-        </Paper>
-      ) : null}
-      <div>
-        <Tooltip
-          title="Quick search"
-          disableHoverListener={searchDisabled}
-          disableFocusListener={searchDisabled}
-          disableTouchListener={searchDisabled}
-        >
-          <div>
-            <IconButton
-              color="primary"
-              onClick={() => {
-                storeSearchTerm("");
-                storeSearchBarDisplayed(!searchBarDisplayed);
-              }}
-              disabled={searchDisabled}
-              data-test="toggle-project-search"
-            >
-              <SearchIcon />
-            </IconButton>
-          </div>
-        </Tooltip>
-      </div>
-    </div>
+    <Searchbar
+      data-test="project-search"
+      searchBarDisplayed={searchBarDisplayed}
+      searchTerm={searchTerm}
+      searchDisabled={searchDisabled}
+      storeSearchBarDisplayed={storeSearchBarDisplayed}
+      storeSearchTerm={storeSearchTerm}
+      autoSearch={true}
+      previewText={strings.common.project_searchtext}
+    />
   );
 };